Fix core tests
