+++ projects/core/src/user/facade/user-consent.service.spec.ts
@@ -1,12 +1,251 @@
-import { TestBed } from '@angular/core/testing';
-
+import { inject, TestBed } from '@angular/core/testing';
+import { Store, StoreModule } from '@ngrx/store';
+import { ConsentTemplate } from '../../model/consent.model';
+import { USERID_CURRENT } from '../../occ/utils/occ-constants';
+import { PROCESS_FEATURE } from '../../process/store/process-state';
+import * as fromProcessReducers from '../../process/store/reducers';
+import * as fromStore from '../store/index';
+import { USER_FEATURE } from '../store/user-state';
 import { UserConsentService } from './user-consent.service';
 
 describe('UserConsentService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
+  let service: UserConsentService;
+  let store: Store<fromStore.UserState>;
 
-  it('should be created', () => {
-    const service: UserConsentService = TestBed.get(UserConsentService);
-    expect(service).toBeTruthy();
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      imports: [
+        StoreModule.forRoot({}),
+        StoreModule.forFeature(USER_FEATURE, fromStore.getReducers()),
+        StoreModule.forFeature(
+          PROCESS_FEATURE,
+          fromProcessReducers.getReducers()
+        ),
+      ],
+      providers: [UserConsentService],
+    });
+
+    store = TestBed.get(Store);
+    spyOn(store, 'dispatch').and.callThrough();
+    service = TestBed.get(UserConsentService);
+  });
+
+  it('should UserConsentService is injected', inject(
+    [UserConsentService],
+    (userConsentService: UserConsentService) => {
+      expect(userConsentService).toBeTruthy();
+    }
+  ));
+
+  const userId = USERID_CURRENT;
+  const consentTemplateListMock: ConsentTemplate[] = [{ id: 'xxx' }];
+
+  describe('load consents', () => {
+    describe('loadConsents', () => {
+      it('should dispatch an action', () => {
+        service.loadConsents();
+        expect(store.dispatch).toHaveBeenCalledWith(
+          new fromStore.LoadUserConsents(userId)
+        );
+      });
+    });
+    describe('getConsents', () => {
+      it('should return the consent template list', () => {
+        store.dispatch(
+          new fromStore.LoadUserConsentsSuccess(consentTemplateListMock)
+        );
+
+        let result: ConsentTemplate[];
+        service
+          .getConsents()
+          .subscribe(consents => (result = consents))
+          .unsubscribe();
+        expect(result).toEqual(consentTemplateListMock);
+      });
+    });
+    describe('getConsentsResultLoading', () => {
+      it('should return the loading flag', () => {
+        store.dispatch(new fromStore.LoadUserConsents(userId));
+
+        let result = false;
+        service
+          .getConsentsResultLoading()
+          .subscribe(loading => (result = loading))
+          .unsubscribe();
+
+        expect(result).toEqual(true);
+      });
+    });
+    describe('getConsentsResultSuccess', () => {
+      it('should return the success flag', () => {
+        store.dispatch(
+          new fromStore.LoadUserConsentsSuccess(consentTemplateListMock)
+        );
+
+        let result = false;
+        service
+          .getConsentsResultSuccess()
+          .subscribe(loading => (result = loading))
+          .unsubscribe();
+
+        expect(result).toEqual(true);
+      });
+    });
+    describe('getConsentsResultError', () => {
+      it('should return the error flag', () => {
+        store.dispatch(new fromStore.LoadUserConsentsFail('an error'));
+
+        let result = false;
+        service
+          .getConsentsResultError()
+          .subscribe(loading => (result = loading))
+          .unsubscribe();
+
+        expect(result).toEqual(true);
+      });
+    });
+    describe('resetConsentsProcessState', () => {
+      it('should dispatch the reset action', () => {
+        service.resetConsentsProcessState();
+        expect(store.dispatch).toHaveBeenCalledWith(
+          new fromStore.ResetLoadUserConsents()
+        );
+      });
+    });
+  });
+
+  describe('give consent', () => {
+    const consentTemplateId = 'templateId';
+    const consentTemplateVersion = 0;
+
+    describe('giveConsent', () => {
+      it('should dispatch an action', () => {
+        service.giveConsent(consentTemplateId, consentTemplateVersion);
+        expect(store.dispatch).toHaveBeenCalledWith(
+          new fromStore.GiveUserConsent({
+            userId,
+            consentTemplateId,
+            consentTemplateVersion,
+          })
+        );
+      });
+    });
+    describe('getGiveConsentResultLoading', () => {
+      it('should return the loading flag', () => {
+        store.dispatch(
+          new fromStore.GiveUserConsent({
+            userId,
+            consentTemplateId,
+            consentTemplateVersion,
+          })
+        );
+
+        let result = false;
+        service
+          .getGiveConsentResultLoading()
+          .subscribe(loading => (result = loading))
+          .unsubscribe();
+
+        expect(result).toEqual(true);
+      });
+    });
+    describe('getGiveConsentResultSuccess', () => {
+      it('should return the success flag', () => {
+        store.dispatch(new fromStore.GiveUserConsentSuccess({}));
+
+        let result = false;
+        service
+          .getGiveConsentResultSuccess()
+          .subscribe(loading => (result = loading))
+          .unsubscribe();
+
+        expect(result).toEqual(true);
+      });
+    });
+    describe('getGiveConsentResultError', () => {
+      it('should return the error flag', () => {
+        store.dispatch(new fromStore.GiveUserConsentFail('an error'));
+
+        let result = false;
+        service
+          .getGiveConsentResultError()
+          .subscribe(loading => (result = loading))
+          .unsubscribe();
+
+        expect(result).toEqual(true);
+      });
+    });
+    describe('resetGiveConsentProcessState', () => {
+      it('should dispatch the reset action', () => {
+        service.resetGiveConsentProcessState();
+        expect(store.dispatch).toHaveBeenCalledWith(
+          new fromStore.ResetGiveUserConsentProcess()
+        );
+      });
+    });
+  });
+
+  describe('withdraw consent', () => {
+    describe('withdrawConsent', () => {
+      it('should dispatch an action', () => {
+        const consentCode = 'xxx';
+        service.withdrawConsent(consentCode);
+        expect(store.dispatch).toHaveBeenCalledWith(
+          new fromStore.WithdrawUserConsent({
+            userId,
+            consentCode,
+          })
+        );
+      });
+    });
+    describe('getWithdrawConsentResultLoading', () => {
+      it('should return the loading flag', () => {
+        store.dispatch(
+          new fromStore.WithdrawUserConsent({ userId, consentCode: 'xxx' })
+        );
+
+        let result = false;
+        service
+          .getWithdrawConsentResultLoading()
+          .subscribe(loading => (result = loading))
+          .unsubscribe();
+
+        expect(result).toEqual(true);
+      });
+    });
+    describe('getWithdrawConsentResultSuccess', () => {
+      it('should return the success flag', () => {
+        store.dispatch(new fromStore.WithdrawUserConsentSuccess());
+
+        let result = false;
+        service
+          .getWithdrawConsentResultSuccess()
+          .subscribe(loading => (result = loading))
+          .unsubscribe();
+
+        expect(result).toEqual(true);
+      });
+    });
+    describe('getWithdrawConsentResultError', () => {
+      it('should return the error flag', () => {
+        store.dispatch(new fromStore.WithdrawUserConsentFail('an error'));
+
+        let result = false;
+        service
+          .getWithdrawConsentResultError()
+          .subscribe(loading => (result = loading))
+          .unsubscribe();
+
+        expect(result).toEqual(true);
+      });
+    });
+    describe('resetWithdrawConsentProcessState', () => {
+      it('should dispatch the reset action', () => {
+        service.resetWithdrawConsentProcessState();
+        expect(store.dispatch).toHaveBeenCalledWith(
+          new fromStore.ResetWithdrawUserConsentProcess()
+        );
+      });
+    });
   });
 });
